Rename 'DevicePermute' to 'DevicePermuteImpl'
