+++ example/37_permute/common.hpp
@@ -15,8 +15,8 @@
 #include <utility>
 
 #include "ck/ck.hpp"
+#include "ck/tensor_operation/gpu/device/impl/device_permute_impl.hpp"
 #include "ck/tensor_operation/gpu/element/binary_element_wise_operation.hpp"
-#include "ck/tensor_operation/gpu/device/device_permute.hpp"
 #include "ck/utility/type.hpp"
 
 #include "ck/library/utility/check_err.hpp"
