+++ example/37_permute/permute_NxHxW_fp16.cpp
@@ -7,7 +7,7 @@ using InDataType  = F16;
 using OutDataType = F16;
 
 // clang-format off
-using DevicePermuteInstance = ck::tensor_operation::device::DevicePermute
+using DevicePermuteInstance = ck::tensor_operation::device::DevicePermuteImpl
 // ######| NumDim|     InData|     OutData| Elementwise| Block|  NPer|  HPer|  WPer|   InBlock|      InBlockTransfer|           InBlockTransfer|       Src|       Dst|             Src|             Dst|
 // ######|       |       Type|        Type|   Operation|  Size| Block| Block| Block| LdsExtraW| ThreadClusterLengths| ThreadClusterArrangeOrder| VectorDim| VectorDim| ScalarPerVector| ScalarPerVector|
 // ######|       |           |            |            |      |      |      |      |          |                     |                          |          |          |                |                |
